+++ DearPyGui/src/Core/mvTextEditor.cpp
@@ -87,7 +87,7 @@ std::string mvTextEditor::GetText(const Coordinates& aStart, const Coordinates& 
 	auto iend = GetCharacterIndex(aEnd);
 	size_t s = 0;
 
-	for (size_t i = lstart; i < lend; i++)
+	for (int i = lstart; i < lend; i++)
 		s += mLines[i].size();
 
 	result.reserve(s + s / 8);
